Refactor model classes to the request packages. this resulted number of file being updated.
