+++ maat-court-data-api/src/main/java/gov/uk/courtdata/prosecutionconcluded/helper/CalculateOutcomeHelper.java
@@ -4,8 +4,8 @@ import gov.uk.courtdata.enums.CrownCourtTrialOutcome;
 import gov.uk.courtdata.enums.PleaTrialOutcome;
 import gov.uk.courtdata.enums.VerdictTrialOutcome;
 import gov.uk.courtdata.exception.ValidationException;
-import gov.uk.courtdata.prosecutionconcluded.listner.request.crowncourt.OffenceSummary;
-import gov.uk.courtdata.prosecutionconcluded.listner.request.crowncourt.ProsecutionConcluded;
+import gov.uk.courtdata.prosecutionconcluded.listner.request.OffenceSummary;
+import gov.uk.courtdata.prosecutionconcluded.listner.request.ProsecutionConcluded;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
