+++ maat-court-data-api/src/main/java/gov/uk/courtdata/prosecutionconcluded/listner/request/Plea.java
@@ -1,4 +1,4 @@
-package gov.uk.courtdata.prosecutionconcluded.listner.request.crowncourt;
+package gov.uk.courtdata.prosecutionconcluded.listner.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
