+++ maat-court-data-api/src/test/java/gov/uk/courtdata/prosecutionconcluded/ProsecutionConcludedListenerTest.java
@@ -3,7 +3,7 @@ package gov.uk.courtdata.prosecutionconcluded;
 import com.google.gson.Gson;
 import gov.uk.courtdata.enums.MessageType;
 import gov.uk.courtdata.prosecutionconcluded.listner.ProsecutionConcludedListener;
-import gov.uk.courtdata.prosecutionconcluded.listner.request.crowncourt.ProsecutionConcluded;
+import gov.uk.courtdata.prosecutionconcluded.listner.request.ProsecutionConcluded;
 import gov.uk.courtdata.prosecutionconcluded.service.ProsecutionConcludedService;
 import gov.uk.courtdata.service.QueueMessageLogService;
